+++ js/widgets/progressbar.js
@@ -1,4 +1,4 @@
-$.widget( "metro.progressBar" , {
+$.widget( "metro.progress" , {
 
     version: "3.0.0",
 
@@ -42,10 +42,10 @@ $.widget( "metro.progressBar" , {
             });
         }
 
-        this.progress(o.value);
+        this.set(o.value);
         this.color(o.color);
 
-        element.data('progressBar', this);
+        element.data('progress', this);
 
     },
 
@@ -67,7 +67,7 @@ $.widget( "metro.progressBar" , {
         o.color = value;
     },
 
-    progress: function(value){
+    set: function(value){
         if (value !== undefined) {
             var element = this.element, o = this.options, colors = this.colorsDim;
             var bar = element.children('.bar:last-child');